+++ genomics_data_index/__init__.py
@@ -1,1 +1,1 @@
-__version__: str = '0.5.0.dev6'
+__version__: str = '0.5.0.dev7'