+++ genomics_data_index/api/query/impl/WrappedSamplesQuery.py
@@ -103,6 +103,8 @@ class WrappedSamplesQuery(SamplesQuery, abc.ABC):
                             kind: str = 'mutations',
                             unit: str = 'percent',
                             category_samples_threshold: int = None,
+                            include_unknown_samples: bool = True, include_unknown_no_present_samples: bool = False,
+                            use_only_samples_in_categories: bool = True,
                             **kwargs) -> pd.DataFrame:
         return self._wrapped_query.features_comparison(sample_categories=sample_categories,
                                                        category_prefixes=category_prefixes,
@@ -110,6 +112,9 @@ class WrappedSamplesQuery(SamplesQuery, abc.ABC):
                                                        kind=kind,
                                                        unit=unit,
                                                        category_samples_threshold=category_samples_threshold,
+                                                       include_unknown_samples=include_unknown_samples,
+                                                       include_unknown_no_present_samples=include_unknown_no_present_samples,
+                                                       use_only_samples_in_categories=use_only_samples_in_categories,
                                                        **kwargs)
 
     def tofeaturesset(self, kind: str = 'mutations', selection: str = 'all',